#2460 Déplacement des fonctions learning_unit_year dans le bon modèle
