+++ base/models/learning_unit_year.py
@@ -23,7 +23,11 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+from itertools import chain
+
 from django.db import models
+
+from base.models.academic_year import current_academic_years
 from osis_common.models.serializable_model import SerializableModel, SerializableModelAdmin
 from base.models.enums import learning_unit_year_activity_status, learning_unit_year_subtypes
 
@@ -101,3 +105,29 @@ def search(academic_year_id=None, acronym=None, learning_container_year_id=None,
         queryset = queryset.filter(activity_status=activity_status)
 
     return queryset.select_related('learning_container_year')
+
+
+def find_all_structure_parents(entities_manager):
+    learning_unit_years_list = list()
+    for entity_manager in entities_manager:
+        learning_unit_years = LearningUnitYear.objects \
+            .filter(structure=entity_manager.structure) \
+            .filter(academic_year=current_academic_years()) \
+            .distinct("structure")
+        for learning_unit_year in learning_unit_years:
+            learning_unit_years = list(chain(learning_unit_years,
+                                             find_all_structure_children(learning_unit_year.structure)))
+        learning_unit_years_list = list(chain(learning_unit_years_list, learning_unit_years))
+    return learning_unit_years_list
+
+
+def find_all_structure_children(structure):
+    learning_unit_years = LearningUnitYear.objects \
+        .filter(structure__part_of=structure) \
+        .filter(academic_year=current_academic_years()) \
+        .distinct("structure")
+    for learning_unit_year in learning_unit_years:
+        if learning_unit_year.structure.part_of:
+            learning_unit_years = list(chain(learning_unit_years,
+                                             find_all_structure_children(learning_unit_year.structure)))
+    return learning_unit_years
